Purge rest of instances in Plutarch.Builtin
