+++ Plutarch/Builtin/Integer.hs
@@ -7,14 +7,23 @@ module Plutarch.Builtin.Integer (
 
   -- * Functions
   pexpModInteger,
+  peqInteger,
+  pleInteger,
+  pltInteger,
+  paddInteger,
+  psubtractInteger,
+  pmultiplyInteger,
+  pconstantInteger,
 ) where
 
 import GHC.Generics (Generic)
 
 -- import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted))
 -- import Plutarch.Internal.Newtype (PlutusTypeNewtype)
+
+import Plutarch.Builtin.Bool (PBool)
 import Plutarch.Builtin.Opaque (POpaque)
-import Plutarch.Internal.Term (S, Term, punsafeBuiltin, (#), (:-->))
+import Plutarch.Internal.Term (S, Term, punsafeBuiltin, punsafeConstantInternal, (#), (:-->))
 import PlutusCore qualified as PLC
 
 {- | A builtin Plutus integer.
@@ -40,3 +49,24 @@ pexpModInteger ::
   forall (s :: S).
   Term s (PInteger :--> PInteger :--> PInteger :--> PInteger)
 pexpModInteger = punsafeBuiltin PLC.ExpModInteger
+
+peqInteger :: forall (s :: S). Term s (PInteger :--> PInteger :--> PBool)
+peqInteger = punsafeBuiltin PLC.EqualsInteger
+
+pleInteger :: forall (s :: S). Term s (PInteger :--> PInteger :--> PBool)
+pleInteger = punsafeBuiltin PLC.LessThanEqualsInteger
+
+pltInteger :: forall (s :: S). Term s (PInteger :--> PInteger :--> PBool)
+pltInteger = punsafeBuiltin PLC.LessThanInteger
+
+paddInteger :: forall (s :: S). Term s (PInteger :--> PInteger :--> PInteger)
+paddInteger = punsafeBuiltin PLC.AddInteger
+
+psubtractInteger :: forall (s :: S). Term s (PInteger :--> PInteger :--> PInteger)
+psubtractInteger = punsafeBuiltin PLC.SubtractInteger
+
+pmultiplyInteger :: forall (s :: S). Term s (PInteger :--> PInteger :--> PInteger)
+pmultiplyInteger = punsafeBuiltin PLC.MultiplyInteger
+
+pconstantInteger :: forall (s :: S). Integer -> Term s PInteger
+pconstantInteger = punsafeConstantInternal . PLC.someValue
