+++ Plutarch/Internal/Show.hs
@@ -1,15 +1,20 @@
+{-# LANGUAGE UndecidableInstances #-}
+
 module Plutarch.Internal.Show (
   PShow (pshow'),
   pshow,
   pshowAndErr,
+  pshowList,
 ) where
 
 import Plutarch.Builtin.Bool
 import Plutarch.Builtin.ByteString
+import Plutarch.Builtin.Data
 import Plutarch.Builtin.Integer
 import Plutarch.Builtin.String
 
 import Data.Char (intToDigit)
+import Data.Kind
 import Data.List.NonEmpty (NonEmpty ((:|)))
 import Data.List.NonEmpty qualified as NE
 import Data.Semigroup (sconcat)
@@ -35,13 +40,17 @@ import Generics.SOP.GGP (gdatatypeInfo)
 
 import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
-import Plutarch.Internal.Lift (pconstant)
+import Plutarch.Internal.IsData
+import Plutarch.Internal.Lift (PlutusRepr, pconstant)
+import Plutarch.Internal.ListLike
 import Plutarch.Internal.Numeric
 import Plutarch.Internal.Ord
 import Plutarch.Internal.Other
 import Plutarch.Internal.PLam
 import Plutarch.Internal.PlutusType
 import Plutarch.Internal.Term
+
+import PlutusCore qualified as PLC
 
 class PShow t where
   -- | Return the string representation of a Plutarch value
@@ -52,9 +61,100 @@ class PShow t where
   default pshow' :: (PGeneric t, PlutusType t, All2 PShow (PCode t)) => Bool -> Term s t -> Term s PString
   pshow' wrap x = gpshow wrap # x
 
+pshowList :: forall list a s. (PShow a, PIsListLike list a) => Term s (list a :--> PString)
+pshowList =
+  phoistAcyclic $
+    plam $ \list ->
+      "[" <> pshowList' @list @a # list <> "]"
+
+pshowList' :: forall list a s. (PShow a, PIsListLike list a) => Term s (list a :--> PString)
+pshowList' =
+  phoistAcyclic $
+    precList
+      ( \self x xs ->
+          pelimList
+            (\_ _ -> pshow x <> ", " <> self # xs)
+            (pshow x)
+            xs
+      )
+      (const "")
+
 -- | Return the string representation of a Plutarch value
 pshow :: PShow a => Term s a -> Term s PString
 pshow = pshow' False
+
+-- | Case matching on bytestring, as if a list.
+pelimBS ::
+  Term
+    s
+    ( PByteString
+        :--> a -- If bytestring is empty
+        :--> (PByte :--> PByteString :--> a) -- If bytestring is non-empty
+        :--> a
+    )
+pelimBS = phoistAcyclic $
+  plam $ \bs z f ->
+    plet (plengthBS # bs) $ \n ->
+      pif (n #== 0) z $
+        plet (pindexBS # bs # 0) $ \x ->
+          plet (psliceBS # 1 # (n - 1) # bs) $ \xs ->
+            f # x # xs
+
+pcase :: PEq a => Term s b -> Term s a -> [(Term s a, Term s b)] -> Term s b
+pcase y x = \case
+  [] -> y
+  ((x', r) : cs) -> pif (x #== x') r $ pcase y x cs
+
+-- | Generic version of `pshow`
+gpshow ::
+  forall a s.
+  (PGeneric a, PlutusType a, All2 PShow (PCode a)) =>
+  Bool ->
+  Term s (a :--> PString)
+gpshow wrap =
+  let constructorNames :: [ConstructorName] =
+        hcollapse $ hmap (K . constructorName) $ constructorInfo $ gdatatypeInfo (Proxy @(a s))
+   in phoistAcyclic $
+        plam $ \x ->
+          pmatch x $ \x' ->
+            productGroup wrap " " $ gpshow' constructorNames (gpfrom x')
+
+-- | Like `gpshow`, but returns the individual parameters list
+gpshow' ::
+  forall a s.
+  All2 PShow a =>
+  [ConstructorName] ->
+  SOP (Term s) a ->
+  NonEmpty (Term s PString)
+gpshow' constructorNames (SOP x) =
+  let cName = constructorNames !! hindex x
+   in pconstant @PString (T.pack cName) :| showSum x
+  where
+    showSum :: NS (NP (Term s)) a -> [Term s PString]
+    showSum =
+      hcollapse . hcmap (Proxy @(All PShow)) showProd
+    showProd :: All PShow xs => NP (Term s) xs -> K [Term s PString] xs
+    showProd =
+      K . hcollapse . hcmap (Proxy @PShow) showTerm
+    showTerm :: forall b. PShow b => Term s b -> K (Term s PString) b
+    showTerm =
+      K . pshow' True
+
+-- | Group parameters list, preparing for final PShow output
+productGroup :: (Monoid a, IsString a) => Bool -> a -> NonEmpty a -> a
+productGroup wrap sep = \case
+  x :| [] -> x
+  xs ->
+    let xs' = sconcat $ NE.intersperse sep xs
+     in if wrap then fromString "(" <> xs' <> fromString ")" else xs'
+
+{- | Causes an error where the input is shown in the message.
+ Works for all types.
+-}
+pshowAndErr :: Term s a -> Term s b
+pshowAndErr x = punsafeCoerce $ pindexBS # punsafeCoerce (pif' # punsafeCoerce x # x # x) # 0
+
+--------------------------------------------------------------------------------
 
 instance PShow PString where
   pshow' _ x = pshowStr # x
@@ -148,73 +248,56 @@ instance PShow PByteString where
               , pconstant @PString $ T.pack [intToDigit x]
               )
 
--- | Case matching on bytestring, as if a list.
-pelimBS ::
-  Term
-    s
-    ( PByteString
-        :--> a -- If bytestring is empty
-        :--> (PByte :--> PByteString :--> a) -- If bytestring is non-empty
-        :--> a
-    )
-pelimBS = phoistAcyclic $
-  plam $ \bs z f ->
-    plet (plengthBS # bs) $ \n ->
-      pif (n #== 0) z $
-        plet (pindexBS # bs # 0) $ \x ->
-          plet (psliceBS # 1 # (n - 1) # bs) $ \xs ->
-            f # x # xs
+instance PShow PData where
+  pshow' b t0 = wrap (go0 # t0)
+    where
+      wrap s = pif (pconstant b) ("(" <> s <> ")") s
+      go0 :: Term s (PData :--> PString)
+      go0 = phoistAcyclic $
+        pfix #$ plam $ \go t ->
+          let pshowConstr pp0 = plet pp0 $ \pp ->
+                "Constr "
+                  <> pshow' False (pfstBuiltin # pp)
+                  <> " "
+                  <> pshowListPString # (pmap # go # (psndBuiltin # pp))
+              pshowMap pplist =
+                "Map " <> pshowListPString # (pmap # pshowPair # pplist)
+              pshowPair = plam $ \pp0 -> plet pp0 $ \pp ->
+                "("
+                  <> (go # (pfstBuiltin # pp))
+                  <> ", "
+                  <> (go # (psndBuiltin # pp))
+                  <> ")"
+              pshowList xs = "List " <> pshowListPString # (pmap # go # xs)
+              pshowListPString = phoistAcyclic $
+                plam $ \plist ->
+                  "["
+                    <> pelimList
+                      ( \x0 xs0 ->
+                          x0 <> (pfoldr' (\x r -> ", " <> x <> r) # ("" :: Term s PString) # xs0)
+                      )
+                      ""
+                      plist
+                    <> "]"
+           in pforce $
+                pchooseData
+                  # t
+                  # pdelay (pshowConstr (pasConstr # t))
+                  # pdelay (pshowMap (pasMap # t))
+                  # pdelay (pshowList (pasList # t))
+                  # pdelay ("I " <> pshow (pasInt # t))
+                  # pdelay ("B " <> pshow (pasByteStr # t))
 
-pcase :: PEq a => Term s b -> Term s a -> [(Term s a, Term s b)] -> Term s b
-pcase y x = \case
-  [] -> y
-  ((x', r) : cs) -> pif (x #== x') r $ pcase y x cs
+instance (PIsData a, PShow a) => PShow (PAsData a) where
+  pshow' w x = pshow' w (pfromData x)
 
--- | Generic version of `pshow`
-gpshow ::
-  forall a s.
-  (PGeneric a, PlutusType a, All2 PShow (PCode a)) =>
-  Bool ->
-  Term s (a :--> PString)
-gpshow wrap =
-  let constructorNames :: [ConstructorName] =
-        hcollapse $ hmap (K . constructorName) $ constructorInfo $ gdatatypeInfo (Proxy @(a s))
-   in phoistAcyclic $
-        plam $ \x ->
-          pmatch x $ \x' ->
-            productGroup wrap " " $ gpshow' constructorNames (gpfrom x')
-
--- | Like `gpshow`, but returns the individual parameters list
-gpshow' ::
-  forall a s.
-  All2 PShow a =>
-  [ConstructorName] ->
-  SOP (Term s) a ->
-  NonEmpty (Term s PString)
-gpshow' constructorNames (SOP x) =
-  let cName = constructorNames !! hindex x
-   in pconstant @PString (T.pack cName) :| showSum x
+instance
+  ( PShow a
+  , PLC.Contains PLC.DefaultUni (PlutusRepr a)
+  ) =>
+  PShow (PBuiltinList a)
   where
-    showSum :: NS (NP (Term s)) a -> [Term s PString]
-    showSum =
-      hcollapse . hcmap (Proxy @(All PShow)) showProd
-    showProd :: All PShow xs => NP (Term s) xs -> K [Term s PString] xs
-    showProd =
-      K . hcollapse . hcmap (Proxy @PShow) showTerm
-    showTerm :: forall b. PShow b => Term s b -> K (Term s PString) b
-    showTerm =
-      K . pshow' True
+  pshow' _ x = pshowList @PBuiltinList @a # x
 
--- | Group parameters list, preparing for final PShow output
-productGroup :: (Monoid a, IsString a) => Bool -> a -> NonEmpty a -> a
-productGroup wrap sep = \case
-  x :| [] -> x
-  xs ->
-    let xs' = sconcat $ NE.intersperse sep xs
-     in if wrap then fromString "(" <> xs' <> fromString ")" else xs'
-
-{- | Causes an error where the input is shown in the message.
- Works for all types.
--}
-pshowAndErr :: Term s a -> Term s b
-pshowAndErr x = punsafeCoerce $ pindexBS # punsafeCoerce (pif' # punsafeCoerce x # x # x) # 0
+instance (PShow a, PShow b) => PShow (PBuiltinPair a b) where
+  pshow' _ pair = "(" <> pshow (pfstBuiltin # pair) <> "," <> pshow (psndBuiltin # pair) <> ")"
