+++ Plutarch/Internal/TryFrom.hs-boot
@@ -1,0 +1,1 @@
+module Plutarch.Internal.TryFrom where
