+++ typos.toml
@@ -8,6 +8,7 @@ Strat = "Strat"
 deriver = "deriver"
 preverse = "preverse"
 intesection = "intesection"
+padd = "padd"
 TIO = "TIO"
 frst = "frst"
 PN = "PN"
